# web/src/lib/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "mtrtk",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/events": {
      "get": {
        "description": "Newest first, by insertion order rather than timestamp.\n\n`ts_utc` comes from the host clock, which steps when NTP or the receiver corrects it, so two\nevents written a second apart can carry timestamps a year apart on a base without an RTC.\nThe id never goes backwards, so it is what \"newest\" means here.\n\nAn unrecognised `level` is a 422, not an empty list: a typo in a filter must not look like a\nclean log.",
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "description": "newest first",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "newest first",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "info, warning or error",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "info",
                    "warning",
                    "error"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "info, warning or error",
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Events Api Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{event_id}/ack": {
      "post": {
        "description": "Mark one event acknowledged. Idempotent, and it publishes nothing: an ack is a UI gesture\nwith no effect on the receiver, and the tab that sent it reads the list back anyway.",
        "operationId": "ack_api_events__event_id__ack_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Ack Api Events  Event Id  Ack Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "no event with that id (retention may already have deleted it)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ack",
        "tags": [
          "events"
        ]
      }
    },
    "/api/history": {
      "get": {
        "description": "Columns-and-rows rather than a list of objects: a day of 1 s samples is 86 400 rows, and\nrepeating the metric names on every one of them would double the response for nothing.",
        "operationId": "history_api_history_get",
        "parameters": [
          {
            "description": "comma-separated metric names",
            "in": "query",
            "name": "metrics",
            "required": true,
            "schema": {
              "description": "comma-separated metric names",
              "title": "Metrics",
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 instant with a timezone, inclusive",
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "description": "ISO-8601 instant with a timezone, inclusive",
              "title": "From",
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 instant with a timezone, exclusive",
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "description": "ISO-8601 instant with a timezone, exclusive",
              "title": "To",
              "type": "string"
            }
          },
          {
            "description": "1m above a six-hour window",
            "in": "query",
            "name": "res",
            "required": false,
            "schema": {
              "default": "auto",
              "description": "1m above a six-hour window",
              "enum": [
                "auto",
                "1s",
                "1m"
              ],
              "title": "Res",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response History Api History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "from/to are not ISO-8601 instants with a timezone, are out of order, span more than the resolution's cap, or a metric is not a column of the chosen table"
          }
        },
        "summary": "History",
        "tags": [
          "history"
        ]
      }
    },
...
    "/api/ntrip/history": {
      "get": {
        "description": "Past connections, newest first - including those made before this daemon started.",
        "operationId": "history_api_ntrip_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response History Api Ntrip History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "ntrip"
        ]
      }
    },
...
  }
}